Remove irrelevant type from content providers.
We won't be categorizing content providers into different types, so
we're removing this field from the interface to keep it clean and less
confusing (we're also deleting the underlying database column and test data).

The majority of the tests touched just get the name of the fixture
updated. We don't want to refer to "portal content provider" and
"organisation content provider" now that we don't have the
portal/organisation/project type.

In one test file the changes are a bit more extensive,
namely test/models/content_provider_test.rb. The tests here were
misusing the fact that the content provider fixtures of various types
just happened to have different states of the 'contact' field (empty,
absent, containing a bare email address, containing both a name and an
email address). The fixtures got new names reflecting their actual
differences and purpose for the tests, which also made possible to
remove the comments that were previously forced to clarify the
intentions of the check.

The development branch was broken with 30 failures and 35 errors when
this commit was made; after the commit is applied, there are 31 failures
and 35 errors, but I believe the new failure was just previously hidden
underneath one of the errors, and should be fixed in the larger effort
of stabilizing the branch and making it green again. For the record, the
new failure is in test/unit/ingestors/ical_ingestor_test.rb.

Commit that added the content provider type back in 2016: 67a63b3d3a74.

Closes #203.