Remove automatic event registration from UI.
This change removes text and icons that  mention things related to
scrapers/ingestors (aka automatic registration of events). We might use
ingestors in the future, and the code around them is extensive, so this
is a superficial UI change rather than a complete removal of ingestors
as a concept.

In particular:
- the helper text in the sidebar on the New Event page got
changed to (1) not mention the automatic registration of events, and (2)
mention the approval process for new events
- the helper text in the sidebar on the Edit Event page got removed (to
  not get into the intricacies of the editing and approval process
depending on the user's role and the event's current approval status)
- the little lock icons are gone from the event form (which is displayed in
  both New and Edit Event pages) — locked fields only make sense in the
context of scraping/ingestors (the lock prevents a value being
overwritten by the ingestor).