Fix the rest of Node ID validation errors in tests
Same as in 3c24b46b8, make event-creating tests a bit more resilient to
changes in the list of mandatory Event fields and clean up after the
PR#196 which made Node ID mandatory but didn't adjust any of the dozens
tests that started failing because of the change. Ideally all these tests
would be rewritten to use well-formed, purpose-appropriate fixtures, but
this is too much effort for too little benefit right now.

I pray for forgiveness of the consistency gods as I also didn't have the
time to unify the naming of the helper variable used across the tests —
it's `@mandatory`, `@mandatory_fields`, `@mandatory_event_fields`,
`@mandatory_fields_of_event` depending on the file.