Fix the test failures from making Node ID mandatory
This commit does more of the same as 3c24b46 and 0306cba15 with two
notable exceptions:

(1) the `:redirect_to_root` `before_action` in the nodes controller had
to be removed to allow any of the nodes controller tests to run. i did
not understand the intention of this redirect from the commit that added
it; my best guess is that we were trying to restrict access, but this is
not how restricting access works. the way it's implemented now it just
breaks the API outright.

(2) a few tests related to locked fields are removed. ideally i
should've done this when i was removing locked fields a couple of months
ago, but back then i couldn't even see these test failures among the
dozens of other, already existing test failures, so i'm cleaning this up
now.