Make language and other fields mandatory for Event
Make language, prerequisites, content providers, cost basis, and target
audience mandatory fields for new Events.

WARNING: while there's a database migration included, and the "pardon
all the old records for their non-validity" workaround creatively called
`after_switch_to_more_mandatory_fields` is there too, the code here
relies heavily on the human agreement to clean up the few records that
we have in the production database before releasing this, as it's just
easier to do that than to e.g. enable the frontend to correctly handle editing
old invalid records or ensuring a valid value for Content Providers for
those Events that don't have it.

Other than that, the change is pretty straightforward. The fixtures and
the magic "all Event's mandatory fields" hashmaps in the tests are
updated to have all the new required values. The tests which get
outdated by this change are deleted (e.g. we don't have to check that an
Event without language set can be saved, with language now being
required it actually should not be saved).

This commit also cleans up an insidious violation of Rails' setup where
the `content_providers` method of `Event` which is supposed to return
a relation, is instead an override that does something else.

Closes #212.